+++ src/Stack/Options/BuildMonoidParser.hs
@@ -36,7 +36,8 @@ import           Stack.Types.ComponentUtils ( unqualCompFromString )
 -- | Parse command line arguments for build configuration.
 buildOptsMonoidParser :: GlobalOptsContext -> Parser BuildOptsMonoid
 buildOptsMonoidParser hide0 = BuildOptsMonoid
-  <$> trace'
+  <$> semaphore
+  <*> trace'
   <*> profile
   <*> noStrip
   <*> libProfiling
@@ -70,7 +71,6 @@ buildOptsMonoidParser hide0 = BuildOptsMonoid
   <*> interleavedOutput
   <*> progressBar
   <*> ddumpDir
-  <*> semaphore
  where
   hideBool = hide0 /= BuildCmdGlobalOpts
   hide :: Mod f a
@@ -247,8 +247,8 @@ buildOptsMonoidParser hide0 = BuildOptsMonoid
     ))
   semaphore = firstBoolFlagsFalse
     "semaphore"
-    "Use Cabal's --semaphore feature to build modules of the same package in \
-    \parallel."
+    "the use of a system semaphore to perform compilation in parallel when \
+    \possible. Supported, by default, by GHC 9.10.1 or later."
     hide
 
 -- | Parser for Cabal verbosity options